digital-credentials/non-secure-contexts.http.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Digital Credentials: non-secure contexts</title>
<link rel="help" href="https://github.com/w3c-fedid/digital-credentials/pull/426" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    assert_equals(typeof window.DigitalCredential, "undefined");
  }, "When in a non-secure context, there is no access to digital credentials.");
</script>
